fixed crash error

+++ src/superbot/actions/commands/recognize_text.js
@@ -2,12 +2,15 @@ import { Resp } from "../../../../lib/superbot/index.js";
 import { createWorker } from "tesseract.js"
 
 export default async function recognize_text_command(cctx, image) {
-    const worker = createWorker();
-    await worker.load();
-    await worker.loadLanguage('rus')
-    await worker.initialize('rus')
-    const { data: { text } } = await worker.recognize(image)
-    worker.terminate()
-    let res = new Resp(`Распознанный текст:\n${text}`)
-    cctx.reply(res);
+    if (image != null && image.match(/.jpg/) != null) {
+        const worker = createWorker();
+        await worker.load();
+        await worker.loadLanguage('rus')
+        await worker.initialize('rus')
+        const { data: { text } } = await worker.recognize(image)
+        worker.terminate()
+        cctx.reply(new Resp(`Распознанный текст:\n${text}`));
+    } else {
+        cctx.reply(new Resp('Проверьте наличие прикреленной картинки к сообщению'))
+    }
 }
